Removed libcanberra from desktop environments to 002-gui .Added xorg-server-xwayland to GNOME stable .Added missing mate-polkit dependencies in current

+++ 003-xfce-4.18/createModule.sh
@@ -66,6 +66,13 @@ currentPackage=atril
 sh $SCRIPTPATH/extras/${currentPackage}/${currentPackage}.SlackBuild || exit 1
 rm -fr $MODULEPATH/${currentPackage}
 
+if [ $SLACKWAREVERSION == "current" ]; then
+	# required by mate-polkit
+	installpkg $MODULEPATH/packages/libappindicator*.txz || exit 1
+	installpkg $MODULEPATH/packages/libdbusmenu*.txz || exit 1
+	installpkg $MODULEPATH/packages/libindicator*.txz || exit 1
+fi
+
 currentPackage=mate-polkit
 sh $SCRIPTPATH/../extras/${currentPackage}/${currentPackage}.SlackBuild || exit 1
 rm -fr $MODULEPATH/${currentPackage}
@@ -115,7 +122,6 @@ ninja -j${NUMBERTHREADS} install || exit 1
 rm -fr $MODULEPATH/${currentPackage}
 
 # required from now on
-installpkg $MODULEPATH/packages/libcanberra*.txz || exit 1
 installpkg $MODULEPATH/packages/libgtop*.txz || exit 1
 
 currentPackage=pavucontrol
@@ -291,11 +297,8 @@ rm usr/bin/vte-*-gtk4
 rm etc/xdg/autostart/blueman.desktop
 rm etc/xdg/autostart/xfce4-clipman-plugin-autostart.desktop
 rm etc/xdg/autostart/xscreensaver.desktop
-rm usr/bin/canberra*
 rm usr/lib${SYSTEMBITS}/girepository-1.0/SoupGNOME*
-rm usr/lib${SYSTEMBITS}/gtk-2.0/modules/libcanberra-gtk-module.*
 rm usr/lib${SYSTEMBITS}/libappindicator.*
-rm usr/lib${SYSTEMBITS}/libcanberra-gtk.*
 rm usr/lib${SYSTEMBITS}/libdbusmenu-gtk.*
 rm usr/lib${SYSTEMBITS}/libindicator.*
 rm usr/lib${SYSTEMBITS}/libkeybinder.*
